# openapi.yaml
openapi: 3.0.3
info:
  title: ACS Multi-Vendor TR-069/CWMP REST API
  version: 1.0.0
  description: |
    REST API internal (Echo v5) untuk Auto Configuration Server (ACS) TR-069/CWMP
    multi-vendor (ZTE, Huawei, FiberHome, Nokia, C-Data, dst). Dikonsumsi oleh
    sistem BSS/OSS, portal NOC, dan integrasi internal lain — **BUKAN** endpoint
    yang dipakai CPE (ONT/ONU/router) itu sendiri.

    **PENTING — beda dengan endpoint CWMP:** perangkat CPE berkomunikasi dengan
    ACS lewat endpoint `/cwmp` (SOAP/XML, protokol TR-069/CWMP standar Broadband
    Forum, sesi stateful) yang **TIDAK** didokumentasikan di file ini sama
    sekali — beda protokol, beda tujuan, beda konsumen. Dokumen ini HANYA
    mencakup `/api/v1/*` (REST + JSON, kecuali `GET /api/v1/metrics` yang
    formatnya teks Prometheus).

    ## Autentikasi

    Hampir seluruh endpoint mewajibkan header `Authorization: Bearer <token>`,
    berupa salah satu dari:
    - **JWT sesi user** — didapat dari `POST /auth/login`, berlaku sesuai masa
      berlaku yang dikonfigurasi (`ACS_JWT_EXPIRY`, default 8 jam di `.env.example`).
    - **API token** — string berawalan `acs_...`, diterbitkan lewat
      `POST /auth/tokens` (ADMIN/SUPERADMIN), tidak pernah kedaluwarsa kecuali
      `expires_at` diisi.

    Empat endpoint publik tanpa autentikasi, semuanya READ-ONLY atau titik masuk
    login (tidak ada endpoint MUTASI tanpa autentikasi): `POST /auth/login`
...
    aman).

...

    ## RBAC

...

...

...
    skema `Task`.

    ## Server

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /tenants:
    post:
      operationId: createTenant
      tags: [Tenants]
      summary: Membuat tenant baru
      description: Superadmin only.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTenantRequest' }
      responses:
        '201':
          description: Tenant dibuat.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tenant' }
        '400':
          description: Payload tidak valid, `code`/`name` kosong, atau `cwmp_inform_password` kurang dari 16 karakter.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Bukan superadmin.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: "`code` tenant sudah dipakai."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    get:
      operationId: listTenants
      tags: [Tenants]
      summary: Daftar seluruh tenant
      description: Superadmin only.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
      responses:
        '200':
          description: Daftar tenant.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TenantListResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Bukan superadmin.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /tenants/current:
    get:
      operationId: getCurrentTenant
      tags: [Tenants]
      summary: Info branding tenant milik actor sendiri
      description: |
        Semua role terautentikasi boleh memanggil (bukan cuma superadmin) —
        dipakai frontend untuk resolve branding (nama, logo, warna aksen) tanpa
        membocorkan field sensitif (`cwmp_inform_username`, kolom audit
        internal). Actor tanpa `tenant_id` (superadmin lintas-tenant)
        mengembalikan `204 No Content`, BUKAN error — frontend disarankan
        fallback ke branding default.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Branding tenant milik actor.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentTenant' }
        '204':
          description: Actor tidak terikat ke tenant mana pun (mis. superadmin global) — tidak ada body.
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /tenants/{id}:
    patch:
      operationId: updateTenant
      tags: [Tenants]
      summary: Update tenant (kode/nama/aktivasi)
      description: |
        Superadmin only. `is_active=false` langsung menolak login/akses SEMUA
        user milik tenant ini pada request berikutnya — termasuk bearer token
        yang sudah diterbitkan sebelumnya (dicek ulang tiap request, bukan
        menunggu token kedaluwarsa). Partial update: field yang tidak dikirim
        (`null`/tidak ada di body) tidak diubah.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/tenantIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateTenantRequest' }
      responses:
        '200':
          description: Tenant setelah diupdate.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tenant' }
        '400':
          description: "`id` path tidak valid atau payload tidak valid."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Bukan superadmin.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Tenant tidak ditemukan.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /tenants/{id}/cwmp-credentials:
    patch:
      operationId: setTenantCWMPCredentials
      tags: [Tenants]
      summary: Set/rotate shared secret Inform CWMP tenant
      description: |
        Superadmin only — kredensial ini divalidasi ACS terhadap Inform dari
        device yang belum tercatat milik tenant ini. Password TIDAK PERNAH
        dikembalikan lagi lewat API manapun setelah di-set (hanya bentuk
        terenkripsi yang tersimpan di DB).
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/tenantIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetTenantCWMPCredentialsRequest' }
      responses:
        '204':
          description: Kredensial berhasil di-set.
        '400':
          description: "`id` tidak valid, payload tidak valid, `username`/`password` kosong, atau password kurang dari 16 karakter."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Bukan superadmin.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Tenant tidak ditemukan.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /tenants/{id}/branding:
    patch:
      operationId: updateTenantBranding
      tags: [Tenants]
      summary: Update branding tenant (nama brand, logo, warna aksen)
      description: |
        SUPERADMIN untuk tenant manapun; ADMIN hanya untuk tenant miliknya
        sendiri (self-service). Ini FULL REPLACE (bukan partial) — field yang
        dikirim `null` eksplisit akan MENGOSONGKAN kolom terkait, beda semantik
        dari `PATCH /tenants/:id` maupun `PATCH /users/:id` yang `null` berarti
        "tidak diubah".
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/tenantIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateTenantBrandingRequest' }
      responses:
        '204':
          description: Branding berhasil diupdate.
        '400':
          description: "`id` tidak valid, atau validasi gagal (`brand_name` >128 karakter, `logo_url` bukan http/https valid atau >512 karakter, `primary_color` bukan format hex `#rrggbb`)."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Role di bawah ADMIN, atau ADMIN mencoba mengubah tenant lain.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /tenants/{id}/task-quota:
    patch:
      operationId: setTenantTaskQuota
      tags: [Tenants]
      summary: Set kuota task queue (`max_pending_tasks`) tenant
      description: |
        Superadmin only — kebijakan platform-level, BUKAN self-service tenant
        (ADMIN tenant sendiri tidak bisa menaikkan kuotanya sendiri lewat
        endpoint ini). `max_pending_tasks: null` berarti tidak dibatasi.
        Kuota ini HANYA membatasi jumlah task berstatus PENDING/QUEUED milik
        tenant, bukan rate limit koneksi/sesi CWMP itu sendiri.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/tenantIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetTenantTaskQuotaRequest' }
      responses:
        '204':
          description: Kuota berhasil diupdate.
        '400':
          description: "`id` tidak valid atau payload tidak valid."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Bukan superadmin.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Tenant tidak ditemukan.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ============================== USERS ==============================
  /users:
    post:
      operationId: createUser
      tags: [Users]
      summary: Membuat user baru
      description: |
        ADMIN/SUPERADMIN. Superadmin bisa membuat user di tenant manapun (atau
        tanpa tenant untuk sesama superadmin); ADMIN non-superadmin hanya bisa
        membuat user di tenant miliknya sendiri, dan TIDAK BOLEH memberikan
        role `SUPERADMIN` ke siapa pun (termasuk lewat variasi casing seperti
        `"superadmin"` — dicek case-insensitive).
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserRequest' }
      responses:
        '201':
          description: User dibuat.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400':
          description: Payload tidak valid, atau `username`/`password` kosong.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Role di bawah ADMIN, ADMIN mencoba membuat user di tenant lain, atau ADMIN mencoba memberi role SUPERADMIN.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: "`username`/`email` sudah dipakai."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    get:
      operationId: listUsers
      tags: [Users]
      summary: Daftar user
      description: |
        ADMIN/SUPERADMIN. Non-superadmin selalu dibatasi ke tenant miliknya
        sendiri — parameter query `tenant_id` diabaikan untuk role ini.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/tenantIdQuery'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
      responses:
        '200':
          description: Daftar user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserListResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Role di bawah ADMIN.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [Users]
      summary: Update data user (nama, email, aktif/nonaktif)
      description: |
        ADMIN/SUPERADMIN. Superadmin bebas ke user manapun; ADMIN dibatasi ke
        user satu tenant dengannya. Guard self-lockout: actor tidak bisa
        menonaktifkan (`is_active=false`) akunnya sendiri lewat endpoint ini.
        Partial update — field yang tidak dikirim tidak diubah.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/userIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateUserRequest' }
      responses:
        '200':
          description: User setelah diupdate.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400':
          description: "`id` tidak valid atau payload tidak valid."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Role di bawah ADMIN, ADMIN mencoba mengubah user tenant lain, atau mencoba menonaktifkan akun sendiri.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: User tidak ditemukan.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    delete:
      operationId: deleteUser
      tags: [Users]
      summary: Soft-delete user
      description: |
        ADMIN/SUPERADMIN, scope sama seperti `PATCH /users/:id`. Guard
        self-lockout: actor tidak bisa menghapus akunnya sendiri.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/userIdPath'
      responses:
        '204':
          description: User berhasil di-soft-delete.
        '400':
          description: "`id` tidak valid."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Role di bawah ADMIN, ADMIN mencoba menghapus user tenant lain, atau mencoba menghapus akun sendiri.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: User tidak ditemukan.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /users/{id}/password:
    patch:
      operationId: resetUserPassword
      tags: [Users]
      summary: Admin mereset password user lain
      description: |
        ADMIN/SUPERADMIN, scope sama seperti `PATCH /users/:id`. BEDA dari
        `PATCH /auth/password` (ganti password sendiri dgn verifikasi password
        lama) — di sini TIDAK PERLU tahu password lama user target. Password
        baru tidak pernah dicatat ke activity log. Sekalian mereset lockout
        brute-force (failed attempts + locked_until) milik user target.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/userIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResetUserPasswordRequest' }
      responses:
        '204':
          description: Password berhasil direset.
        '400':
          description: "`id` tidak valid, payload tidak valid, `new_password` kosong atau kurang dari 8 karakter."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Role di bawah ADMIN, atau ADMIN mencoba mereset password user tenant lain.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: User tidak ditemukan.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /users/{id}/roles:
    patch:
      operationId: replaceUserRoles
      tags: [Users]
      summary: Ganti seluruh role user (full replace)
      description: |
        ADMIN/SUPERADMIN, scope sama seperti `PATCH /users/:id`. FULL REPLACE —
        role yang tidak disebut di `role_codes` akan DICABUT. Dua guard
        privilege-escalation/self-lockout: (1) ADMIN non-superadmin tidak boleh
        memberikan role `SUPERADMIN` ke siapa pun; (2) actor (ADMIN) tidak boleh
        mencabut role `ADMIN` dari akunnya sendiri.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/userIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReplaceUserRolesRequest' }
      responses:
        '200':
          description: User dengan daftar role terbaru.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400':
          description: "`id` tidak valid, payload tidak valid, atau ada `role_code` yang tidak dikenal."
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Tidak terautentikasi.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Role di bawah ADMIN, ADMIN mencoba mengubah user tenant lain, ADMIN mencoba memberi role SUPERADMIN, atau mencoba mencabut role ADMIN dari diri sendiri.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: User tidak ditemukan.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ============================== DEVICES ==============================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT sesi user (dari `POST /auth/login`) ATAU API token statis
        berawalan `acs_...` (dari `POST /auth/tokens`) — keduanya dikirim
        dengan cara yang sama lewat header `Authorization: Bearer <token>`.
  parameters:
    page:
      name: page
      in: query
      description: Halaman (1-based). Default 1.
      schema: { type: integer, minimum: 1, default: 1 }
    pageSize:
      name: page_size
      in: query
      description: Ukuran halaman. Default 50, maksimum 200 (nilai di luar rentang ini jatuh ke default 50).
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
    tenantIdPath:
      name: id
      in: path
      required: true
      description: ID numerik internal tenant.
      schema: { type: integer, format: int64 }
...
  schemas:
...
    ErrorResponse:
      type: object
      description: |
        Bentuk error SERAGAM untuk seluruh endpoint (`echo.NewHTTPError`) —
        selalu `{"message": "..."}`, tidak ada kode error mesin terpisah dari
        `message` teks bebas saat ini (perhatikan: teks pesan sebagian besar
        Bahasa Indonesia, konsisten dengan gaya pesan aplikasi ini).
      properties:
        message:
          type: string
          example: "domain: data tidak ditemukan"
      required: [message]
...
    Tenant:
      type: object
      properties:
        id: { type: integer, format: int64 }
        tenant_uuid: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        is_active: { type: boolean }
        cwmp_inform_username:
          type: string
          nullable: true
          description: "Separuh dari shared secret Inform CWMP tenant ini. Separuh lainnya (password) TIDAK PERNAH diekspos di response manapun."
        brand_name: { type: string, nullable: true }
        logo_url: { type: string, nullable: true }
        primary_color: { type: string, nullable: true, example: "#1a73e8" }
        max_pending_tasks: { type: integer, nullable: true, description: "NULL berarti tidak dibatasi." }
      allOf:
        - $ref: '#/components/schemas/Audit'
      required: [id, tenant_uuid, code, name, is_active]
    TenantListResponse:
      type: object
      properties:
        data:
          type: array
          nullable: true
          items: { $ref: '#/components/schemas/Tenant' }
        total: { type: integer }
      required: [data, total]
...
    CreateTenantRequest:
      type: object
      description: "Batas `maxLength` di bawah ditegakkan server (`delivery/http/validate.go`) dan sama persis dengan lebar kolom `schema.sql` — melebihi batas dijawab `400`, bukan `500`."
      properties:
        code: { type: string, maxLength: 32 }
        name: { type: string, maxLength: 128 }
        cwmp_inform_username: { type: string, nullable: true, maxLength: 128 }
        cwmp_inform_password:
          type: string
          format: password
          nullable: true
          minLength: 16
          description: "Tidak dibatasi panjang maksimum di server (disimpan terenkripsi sebagai VARBINARY, bukan VARCHAR)."
      required: [code, name]
...
    SetTenantTaskQuotaRequest:
      type: object
      properties:
        max_pending_tasks:
          type: integer
          nullable: true
          description: "NULL berarti tidak dibatasi."
...